requirements-dev.txt
-r requirements.txt
django-debug-toolbar==1.9.1
django-silk==2.0.0
wheel==0.31.0